transform_doc/requirements.txt
# Основные библиотеки для работы с документами
python-docx==1.1.0

# Библиотеки для работы с LLM
requests>=2.32.5
python-dotenv>=0.19.0

# Библиотеки для логирования
# (встроенные в Python: logging, logging.handlers)

# Дополнительные зависимости для python-docx
lxml>=3.1.0
typing-extensions>=4.0.0